Rename `app-util` -> `sys-util`.
