+++ src/builtin-services/export/RequestLogger.js
@@ -5,7 +5,7 @@ import * as fs from 'node:fs/promises';
 
 import { FileServiceConfig } from '@this/sys-config';
 import { BaseService } from '@this/sys-framework';
-import { Rotator } from '@this/app-util';
+import { Rotator } from '@this/sys-util';
 import { Moment } from '@this/data-values';
 import { IntfLogger } from '@this/loggy';
 import { IntfRequestLogger } from '@this/network-protocol';
