+++ src/builtin-services/export/SystemLogger.js
@@ -5,7 +5,7 @@ import * as timers from 'node:timers/promises';
 
 import { FileServiceConfig } from '@this/sys-config';
 import { BaseService } from '@this/sys-framework';
-import { Rotator } from '@this/app-util';
+import { Rotator } from '@this/sys-util';
 import { EventTracker, LinkedEvent } from '@this/async';
 import { IntfLogger, Loggy, TextFileSink } from '@this/loggy';
 import { MustBe } from '@this/typey';
