+++ src/main-lactoserv/private/UsualSystem.js
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Warehouse } from '@this/sys-framework';
-import { BaseSystem } from '@this/app-util';
+import { BaseSystem } from '@this/sys-util';
 
 import { MainArgs } from '#p/MainArgs';
 import { ThisModule } from '#p/ThisModule';
