+++ src/sys-util/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/app-util",
+  "name": "@this/sys-util",
   "version": "0.0.1",
   "type": "module",
   "private": true,
