+++ src/sys-util/private/BaseFilePreserver.js
